+++ formio.dart
@@ -1,4 +1,0 @@
-library formio;
-
-export 'packages/flutter_formio/lib/flutter_formio.dart';
-export 'packages/formio_api/lib/formio_api.dart';