+++ packages/flutter_formio/lib/src/widgets/components/container_component.dart
@@ -6,7 +6,7 @@
 library;
 
 import 'package:flutter/material.dart';
-import 'package:flutter_formio/flutter_formio.dart';
+import 'package:formio/flutter_formio.dart';
 
 class ContainerComponent extends StatelessWidget {
   /// The Form.io component definition for the container.