+++ packages/flutter_formio/lib/src/widgets/components/data_grid_component.dart
@@ -7,7 +7,7 @@
 library;
 
 import 'package:flutter/material.dart';
-import 'package:flutter_formio/flutter_formio.dart';
+import 'package:formio/flutter_formio.dart';
 
 class DataGridComponent extends StatefulWidget {
   /// The Form.io component definition.