+++ packages/flutter_formio/lib/src/widgets/components/fieldset_component.dart
@@ -5,7 +5,7 @@
 library;
 
 import 'package:flutter/material.dart';
-import 'package:flutter_formio/flutter_formio.dart';
+import 'package:formio/flutter_formio.dart';
 
 class FieldSetComponent extends StatelessWidget {
   /// The Form.io fieldset component.