+++ packages/flutter_formio/test/component_factory_plugin_test.dart
@@ -1,9 +1,9 @@
 // ignore_for_file: depend_on_referenced_packages
 
 import 'package:flutter/material.dart';
+import 'package:formio/flutter_formio.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../formio.dart';
 
 void main() {
   group('ComponentFactory Plugin System', () {